Rename custom to app
