+++ Source/ApiTemplate/Source/ApiTemplate/AppServiceCollectionExtensions.cs
@@ -1,27 +1,21 @@
-namespace GraphQLTemplate;
+namespace ApiTemplate;
 
-using Boxed.AspNetCore;
-using GraphQLTemplate.ConfigureOptions;
-#if ((HealthCheck && Redis) || Redis || PersistedQueries || Subscriptions)
-using GraphQLTemplate.Constants;
+using ApiTemplate.ConfigureOptions;
+#if (HealthCheck && Redis)
+using ApiTemplate.Constants;
 #endif
-using GraphQLTemplate.Directives;
-using GraphQLTemplate.Options;
-using GraphQLTemplate.Types;
-using HotChocolate.Execution.Options;
+using ApiTemplate.Options;
+using Boxed.AspNetCore;
 #if (!ForwardedHeaders && HostFiltering)
 using Microsoft.AspNetCore.HostFiltering;
 #endif
 using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.Options;
-#if Redis
-using StackExchange.Redis;
-#endif
 
 /// <summary>
 /// <see cref="IServiceCollection"/> extension methods which extend ASP.NET Core services.
 /// </summary>
-internal static class CustomServiceCollectionExtensions
+internal static class AppServiceCollectionExtensions
 {
     /// <summary>
     /// Configures the settings by binding the contents of the appsettings.json file to the specified Plain Old CLR
@@ -29,7 +23,7 @@ internal static class CustomServiceCollectionExtensions
     /// </summary>
     /// <param name="services">The services.</param>
     /// <param name="configuration">The configuration.</param>
-    /// <returns>The services with caching services added.</returns>
+    /// <returns>The services with options services added.</returns>
     public static IServiceCollection AddCustomOptions(
         this IServiceCollection services,
         IConfiguration configuration) =>
@@ -51,19 +45,19 @@ internal static class CustomServiceCollectionExtensions
 #elif HostFiltering
             .ConfigureAndValidateSingleton<HostFilteringOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.HostFiltering)))
 #endif
-            .ConfigureAndValidateSingleton<GraphQLOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.GraphQL)))
-            .ConfigureAndValidateSingleton<RequestExecutorOptions>(
-                configuration.GetRequiredSection(nameof(ApplicationOptions.GraphQL)).GetRequiredSection(nameof(GraphQLOptions.Request)))
+            .ConfigureAndValidateSingleton<HostOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.Host)))
 #if Redis
             .ConfigureAndValidateSingleton<RedisOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.Redis)))
 #endif
-            .ConfigureAndValidateSingleton<HostOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.Host)))
             .ConfigureAndValidateSingleton<KestrelServerOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.Kestrel)));
 
     public static IServiceCollection AddCustomConfigureOptions(this IServiceCollection services) =>
         services
-#if Authorization
-            .ConfigureOptions<ConfigureAuthorizationOptions>()
+#if Versioning
+            .ConfigureOptions<ConfigureApiVersioningOptions>()
+#endif
+#if Controllers
+            .ConfigureOptions<ConfigureMvcOptions>()
 #endif
 #if CORS
             .ConfigureOptions<ConfigureCorsOptions>()
@@ -71,7 +65,8 @@ internal static class CustomServiceCollectionExtensions
 #if HstsPreload
             .ConfigureOptions<ConfigureHstsOptions>()
 #endif
-#if DistributedCacheRedis
+            .ConfigureOptions<ConfigureJsonOptions>()
+#if Redis
             .ConfigureOptions<ConfigureRedisCacheOptions>()
 #endif
 #if Serilog
@@ -81,6 +76,10 @@ internal static class CustomServiceCollectionExtensions
             .ConfigureOptions<ConfigureResponseCompressionOptions>()
 #endif
             .ConfigureOptions<ConfigureRouteOptions>()
+#if Swagger
+            .ConfigureOptions<ConfigureSwaggerGenOptions>()
+            .ConfigureOptions<ConfigureSwaggerUIOptions>()
+#endif
             .ConfigureOptions<ConfigureStaticFileOptions>();
 #if HealthCheck
 
@@ -102,68 +101,4 @@ internal static class CustomServiceCollectionExtensions
 #endif
             .Services;
 #endif
-#if Redis
-
-    public static IServiceCollection AddCustomRedis(
-        this IServiceCollection services,
-        IWebHostEnvironment webHostEnvironment,
-        IConfiguration configuration) =>
-        services.AddIf(
-            !webHostEnvironment.IsEnvironment(EnvironmentName.Test),
-            x => x.AddSingleton<IConnectionMultiplexer>(
-                ConnectionMultiplexer.Connect(
-                     configuration
-                        .GetRequiredSection(nameof(ApplicationOptions.Redis))
-                        .Get<RedisOptions>()
-                        .ConfigurationOptions)));
-#endif
-
-    public static IServiceCollection AddCustomGraphQL(
-        this IServiceCollection services,
-#if (PersistedQueries || Subscriptions)
-        IWebHostEnvironment webHostEnvironment,
-#endif
-        IConfiguration configuration)
-    {
-        var graphQLOptions = configuration.GetRequiredSection(nameof(ApplicationOptions.GraphQL)).Get<GraphQLOptions>();
-        return services
-            .AddGraphQLServer()
-#if OpenTelemetry
-            .AddInstrumentation()
-#endif
-            .InitializeOnStartup()
-            .AddFiltering()
-            .AddSorting()
-            .AddGlobalObjectIdentification()
-            .AddQueryFieldToMutationPayloads()
-            .AddApolloTracing()
-#if Authorization
-            .AddAuthorization()
-#endif
-#if PersistedQueries
-            .UseAutomaticPersistedQueryPipeline()
-            .AddIfElse(
-                webHostEnvironment.IsEnvironment(EnvironmentName.Test),
-                x => x.AddInMemoryQueryStorage(),
-                x => x.AddRedisQueryStorage())
-#endif
-#if Subscriptions
-            .AddIfElse(
-                webHostEnvironment.IsEnvironment(EnvironmentName.Test),
-                x => x.AddInMemorySubscriptions(),
-                x => x.AddRedisSubscriptions())
-#endif
-            .SetSchema<MainSchema>()
-            .AddDirectiveType<UpperDirectiveType>()
-            .AddDirectiveType<LowerDirectiveType>()
-            .AddDirectiveType<IncludeDirectiveType>()
-            .AddDirectiveType<SkipDirectiveType>()
-            .AddGraphQLTemplateTypes()
-            .TrimTypes()
-            .ModifyOptions(options => options.UseXmlDocumentation = false)
-            .AddMaxExecutionDepthRule(graphQLOptions.MaxAllowedExecutionDepth)
-            .SetPagingOptions(graphQLOptions.Paging)
-            .SetRequestOptions(() => graphQLOptions.Request)
-            .Services;
-    }
 }
