+++ Source/GraphQLTemplate/Source/GraphQLTemplate/AppServiceCollectionExtensions.cs
@@ -1,21 +1,27 @@
-namespace ApiTemplate;
+namespace GraphQLTemplate;
 
-using ApiTemplate.ConfigureOptions;
-#if (HealthCheck && Redis)
-using ApiTemplate.Constants;
-#endif
-using ApiTemplate.Options;
 using Boxed.AspNetCore;
+using GraphQLTemplate.ConfigureOptions;
+#if ((HealthCheck && Redis) || Redis || PersistedQueries || Subscriptions)
+using GraphQLTemplate.Constants;
+#endif
+using GraphQLTemplate.Directives;
+using GraphQLTemplate.Options;
+using GraphQLTemplate.Types;
+using HotChocolate.Execution.Options;
 #if (!ForwardedHeaders && HostFiltering)
 using Microsoft.AspNetCore.HostFiltering;
 #endif
 using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.Options;
+#if Redis
+using StackExchange.Redis;
+#endif
 
 /// <summary>
 /// <see cref="IServiceCollection"/> extension methods which extend ASP.NET Core services.
 /// </summary>
-internal static class CustomServiceCollectionExtensions
+internal static class AppServiceCollectionExtensions
 {
     /// <summary>
     /// Configures the settings by binding the contents of the appsettings.json file to the specified Plain Old CLR
@@ -23,7 +29,7 @@ internal static class CustomServiceCollectionExtensions
     /// </summary>
     /// <param name="services">The services.</param>
     /// <param name="configuration">The configuration.</param>
-    /// <returns>The services with options services added.</returns>
+    /// <returns>The services with caching services added.</returns>
     public static IServiceCollection AddCustomOptions(
         this IServiceCollection services,
         IConfiguration configuration) =>
@@ -45,19 +51,19 @@ internal static class CustomServiceCollectionExtensions
 #elif HostFiltering
             .ConfigureAndValidateSingleton<HostFilteringOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.HostFiltering)))
 #endif
-            .ConfigureAndValidateSingleton<HostOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.Host)))
+            .ConfigureAndValidateSingleton<GraphQLOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.GraphQL)))
+            .ConfigureAndValidateSingleton<RequestExecutorOptions>(
+                configuration.GetRequiredSection(nameof(ApplicationOptions.GraphQL)).GetRequiredSection(nameof(GraphQLOptions.Request)))
 #if Redis
             .ConfigureAndValidateSingleton<RedisOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.Redis)))
 #endif
+            .ConfigureAndValidateSingleton<HostOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.Host)))
             .ConfigureAndValidateSingleton<KestrelServerOptions>(configuration.GetRequiredSection(nameof(ApplicationOptions.Kestrel)));
 
     public static IServiceCollection AddCustomConfigureOptions(this IServiceCollection services) =>
         services
-#if Versioning
-            .ConfigureOptions<ConfigureApiVersioningOptions>()
-#endif
-#if Controllers
-            .ConfigureOptions<ConfigureMvcOptions>()
+#if Authorization
+            .ConfigureOptions<ConfigureAuthorizationOptions>()
 #endif
 #if CORS
             .ConfigureOptions<ConfigureCorsOptions>()
@@ -65,8 +71,7 @@ internal static class CustomServiceCollectionExtensions
 #if HstsPreload
             .ConfigureOptions<ConfigureHstsOptions>()
 #endif
-            .ConfigureOptions<ConfigureJsonOptions>()
-#if Redis
+#if DistributedCacheRedis
             .ConfigureOptions<ConfigureRedisCacheOptions>()
 #endif
 #if Serilog
@@ -76,10 +81,6 @@ internal static class CustomServiceCollectionExtensions
             .ConfigureOptions<ConfigureResponseCompressionOptions>()
 #endif
             .ConfigureOptions<ConfigureRouteOptions>()
-#if Swagger
-            .ConfigureOptions<ConfigureSwaggerGenOptions>()
-            .ConfigureOptions<ConfigureSwaggerUIOptions>()
-#endif
             .ConfigureOptions<ConfigureStaticFileOptions>();
 #if HealthCheck
 
@@ -101,4 +102,68 @@ internal static class CustomServiceCollectionExtensions
 #endif
             .Services;
 #endif
+#if Redis
+
+    public static IServiceCollection AddCustomRedis(
+        this IServiceCollection services,
+        IWebHostEnvironment webHostEnvironment,
+        IConfiguration configuration) =>
+        services.AddIf(
+            !webHostEnvironment.IsEnvironment(EnvironmentName.Test),
+            x => x.AddSingleton<IConnectionMultiplexer>(
+                ConnectionMultiplexer.Connect(
+                     configuration
+                        .GetRequiredSection(nameof(ApplicationOptions.Redis))
+                        .Get<RedisOptions>()
+                        .ConfigurationOptions)));
+#endif
+
+    public static IServiceCollection AddCustomGraphQL(
+        this IServiceCollection services,
+#if (PersistedQueries || Subscriptions)
+        IWebHostEnvironment webHostEnvironment,
+#endif
+        IConfiguration configuration)
+    {
+        var graphQLOptions = configuration.GetRequiredSection(nameof(ApplicationOptions.GraphQL)).Get<GraphQLOptions>();
+        return services
+            .AddGraphQLServer()
+#if OpenTelemetry
+            .AddInstrumentation()
+#endif
+            .InitializeOnStartup()
+            .AddFiltering()
+            .AddSorting()
+            .AddGlobalObjectIdentification()
+            .AddQueryFieldToMutationPayloads()
+            .AddApolloTracing()
+#if Authorization
+            .AddAuthorization()
+#endif
+#if PersistedQueries
+            .UseAutomaticPersistedQueryPipeline()
+            .AddIfElse(
+                webHostEnvironment.IsEnvironment(EnvironmentName.Test),
+                x => x.AddInMemoryQueryStorage(),
+                x => x.AddRedisQueryStorage())
+#endif
+#if Subscriptions
+            .AddIfElse(
+                webHostEnvironment.IsEnvironment(EnvironmentName.Test),
+                x => x.AddInMemorySubscriptions(),
+                x => x.AddRedisSubscriptions())
+#endif
+            .SetSchema<MainSchema>()
+            .AddDirectiveType<UpperDirectiveType>()
+            .AddDirectiveType<LowerDirectiveType>()
+            .AddDirectiveType<IncludeDirectiveType>()
+            .AddDirectiveType<SkipDirectiveType>()
+            .AddGraphQLTemplateTypes()
+            .TrimTypes()
+            .ModifyOptions(options => options.UseXmlDocumentation = false)
+            .AddMaxExecutionDepthRule(graphQLOptions.MaxAllowedExecutionDepth)
+            .SetPagingOptions(graphQLOptions.Paging)
+            .SetRequestOptions(() => graphQLOptions.Request)
+            .Services;
+    }
 }
