+++ Source/GraphQLTemplate/Tests/GraphQLTemplate.IntegrationTest/AppWebApplicationFactory.cs
@@ -1,14 +1,13 @@
-namespace ApiTemplate.IntegrationTest;
+namespace GraphQLTemplate.IntegrationTest;
 
-using ApiTemplate.Options;
-#if Controllers
-using ApiTemplate.Repositories;
-using ApiTemplate.Services;
-#endif
+using System;
+using System.Net.Http;
+using GraphQLTemplate.Options;
+using GraphQLTemplate.Services;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
-#if Controllers
+using Microsoft.Extensions.DependencyInjection;
 using Moq;
-#endif
 #if Serilog
 using Serilog;
 using Serilog.Events;
@@ -17,11 +16,11 @@ using Serilog.Events;
 using Xunit.Abstractions;
 #endif
 
-public class CustomWebApplicationFactory<TEntryPoint> : WebApplicationFactory<TEntryPoint>
+public class AppWebApplicationFactory<TEntryPoint> : WebApplicationFactory<TEntryPoint>
     where TEntryPoint : class
 {
 #if Serilog
-    public CustomWebApplicationFactory(ITestOutputHelper testOutputHelper)
+    public AppWebApplicationFactory(ITestOutputHelper testOutputHelper)
     {
         this.ClientOptions.AllowAutoRedirect = false;
 #if HttpsEverywhere
@@ -45,14 +44,10 @@ public class CustomWebApplicationFactory<TEntryPoint> : WebApplicationFactory<TE
 
     public ApplicationOptions ApplicationOptions { get; private set; } = default!;
 
-#if Controllers
-    public Mock<ICarRepository> CarRepositoryMock { get; } = new Mock<ICarRepository>(MockBehavior.Strict);
-
     public Mock<IClockService> ClockServiceMock { get; } = new Mock<IClockService>(MockBehavior.Strict);
 
-    public void VerifyAllMocks() => Mock.VerifyAll(this.CarRepositoryMock, this.ClockServiceMock);
+    public void VerifyAllMocks() => Mock.VerifyAll(this.ClockServiceMock);
 
-#endif
     protected override void ConfigureClient(HttpClient client)
     {
         using (var serviceScope = this.Services.CreateScope())
@@ -66,30 +61,15 @@ public class CustomWebApplicationFactory<TEntryPoint> : WebApplicationFactory<TE
 
     protected override void ConfigureWebHost(IWebHostBuilder builder) =>
         builder
-#if (DistributedCacheRedis || Controllers)
-            .ConfigureServices(this.ConfigureServices)
-#endif
-            .UseEnvironment(Constants.EnvironmentName.Test);
-#if (DistributedCacheRedis && Controllers)
+            .UseEnvironment(Constants.EnvironmentName.Test)
+            .ConfigureServices(this.ConfigureServices);
 
     protected virtual void ConfigureServices(IServiceCollection services) =>
         services
+#if DistributedCacheRedis
             .AddDistributedMemoryCache()
-            .AddSingleton(this.CarRepositoryMock.Object)
-            .AddSingleton(this.ClockServiceMock.Object);
-#elif DistributedCacheRedis
-
-    protected virtual void ConfigureServices(IServiceCollection services) =>
-        services
-            .AddDistributedMemoryCache();
-#elif Controllers
-
-    protected virtual void ConfigureServices(IServiceCollection services) =>
-        services
-            .AddSingleton(this.CarRepositoryMock.Object)
-            .AddSingleton(this.ClockServiceMock.Object);
 #endif
-#if Controllers
+            .AddSingleton(this.ClockServiceMock.Object);
 
     protected override void Dispose(bool disposing)
     {
@@ -100,5 +80,4 @@ public class CustomWebApplicationFactory<TEntryPoint> : WebApplicationFactory<TE
 
         base.Dispose(disposing);
     }
-#endif
 }
